Unit test the CLI interface.
Moves the CLI entrypoint next to Diff/PatchOperations so we can use package access rules to peek into the internals for the unit tests. We are mostly just asserting that the CLI can parse the various input/output modes. The others are a bonus.

Fixes a bug with the recent `archive-patches` argument. Which had an overlapping shorhand argument `M` with the diff variant. This caused it to be unusable, along with the blatantly wrong `availableIf`.