Rebuild all unit tests to actually test functionality.
The test infrastructure is now in a much better state. Previously all these tests tried to test the functionality difference between Archives and Folders. 

Now that we have a unified IO system, with test to assert that there is no difference. We can focus these tests on the actual diff/patch handling.